requirements.txt
streamlit
pandas
polars
numpy
matplotlib
openpyxl
python-docx
SQLAlchemy
plotly
scikit-learn
pytest
scipy
joblib
shap
boto3
google-cloud-storage
azure-storage-blob
dask[complete]
fpdf2
streamlit-authenticator
pytest-cov
kaleido
pydantic
ruff
supabase
python-dotenv